Resolving issues with video sharing
