--- src/app/core/components/logo/logo.component.scss
@@ -1,1 +1,0 @@
-@import '../../../../variables';
