+++ src/app/core/components/nav/nav-menu-item/nav-menu-item.component.scss
@@ -1,4 +1,6 @@
-@import '../../../../../variables';
+//@import '../../../../../variables';
+
+$primaryColor:  #607d8b;
 
 .sidenav-item {
     position: relative;
