+++ src/app/core/components/record-search/record-search.component.html
@@ -1,74 +1,18 @@
-<!-- <mat-form-field class="er-case-search-box" [@expandSearch]="searchShowing ? 'open' : 'closed'"> -->
-   <!--<mat-form-field class="er-case-search-box" >
+<app-search-field (searchString)="onSearchQuery($event)"></app-search-field>
 
-    <span matPrefix (click)="executeSearch()">
-        <mat-icon style="cursor: pointer;" matSuffix>search</mat-icon>
-    </span>
-    <input class="er-case-search-box" autocomplete="off" [(ngModel)]="search.searchString" type="text" (keyup.enter)="executeSearch()" matInput>
-    <mat-icon style="cursor: pointer;" matSuffix (click)="toggleSearchBox()">keyboard_arrow_down</mat-icon>
-</mat-form-field>
+<div fxLayout="row wrap" fxLayoutGap="20px" fxLayoutAlign="start" class="search-results">
 
+    <div *ngIf="searchResults$ | async as searchResults;">
 
-<div [@expandSearchForm]="searchShowing ? 'open' : 'closed'"
-    class="search-form" [formGroup]="searchForm">
+        <div *ngIf="searchResults.length === 0;">No results</div>
 
-        <div formArrayName="searchRows" *ngFor="let item of searchRows?.controls; let i = index;">
-            <div [formGroupName]="i">
-
-                <div fxLayout="row" fxLayoutGap="20px" fxLayoutAlign="space-between center">
-                    <div fxFlex="5">
-                        <mat-icon style="cursor: pointer; margin-left:15px;" (click)="removeRow(i)">
-                            remove_circle_outline</mat-icon>
-                    </div>
-
-                    <div fxLayout="row" fxLayout.xs="column" fxLayoutGap="20px" fxLayoutGap.xs="0px" fxLayoutAlign="center center" fxFlex>
-
-                    <div fxFlex>
-
-                        <mat-form-field>
-                            <mat-label>Search field</mat-label>
-                            <mat-select formControlName="searchField">
-                                <mat-option *ngFor="let option of options" #id{{i}} [value]="option.id">
-                                    {{option.name}}
-                                </mat-option>
-                            </mat-select>
-                        </mat-form-field>
-
-
-                        </div>
-                        <div fxFlex>
-                            <mat-form-field>
-                                <input formControlName="searchTerm" autocomplete="off" matInput>
-                            </mat-form-field>
-                        </div>
-                    </div>
-                </div>
-
-            </div>
-
-        </div>
-
-        <mat-icon class="add-search-row" (click)="addRow()">add_circle_outline</mat-icon>
-
-        <button mat-raised-button class="er-search-button" (click)="executeSearch()">
-            Search
-        </button>
-    </div>
-
-<div>-->
-
-    <app-search-box (onSearchQuery)="searchQuery")></app-search-box>
-
-    <div fxLayout="row wrap" fxLayoutGap="20px" fxLayoutAlign="start" class="search-results">
-
-        <div *ngIf="(searchResults$ | async)?.length === 0 ;">No results</div>
-
-        <div *ngFor="let record of searchResults$ | async">
+        <div *ngFor="let record of searchResults">
 
             <search-result-card class="emergency-card" (openEmergencyCase)="openCase($event)" [record]="record"></search-result-card>
 
         </div>
 
     </div>
+
 
 </div>
