+++ src/app/core/components/record-search/record-search.component.scss
@@ -1,46 +1,6 @@
 $standard-width: 175px;
 $standard-height: 48px;
 
-.er-case-search-box {
-    max-width: $standard-width * 3 !important;
-    width: 97%;
-    padding-bottom: 1px;
-}
-
-.search-form {
-    width: 97%;
-    max-width: $standard-width * 3;
-    border: 1px rgba(0, 0, 0, 0.36) solid;
-    border-radius: 4px;
-    position: relative;
-    background: #ffffff;
-    padding-bottom: 40px;
-    margin-top: -17px;
-}
-
-.er-search-button {
-    position: absolute;
-    bottom: 0px;
-    right: 0px;
-    margin-bottom: 5px;
-    margin-right: 5px;
-}
-
-.search-row {
-    margin: 15px;
-    height: 48px;
-}
-
-.add-search-row {
-    margin-left: 15px;
-    margin-right: 15px;
-    height: 48px;
-    vertical-align: middle;
-    position: relative;
-    top: 15px;
-    cursor: pointer;
-}
-
 ::ng-deep .emergency-card {
     max-width: 220px !important;
     height: fit-content;
