--- src/app/core/components/search-box/search-box.component.scss
@@ -1,3 +1,0 @@
-.expansion-panel-margin {
-    margin-top: 5px;
-}
