--- src/app/core/components/search-box/search-box.module.ts
@@ -1,12 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common'; 
-
-@NgModule({
-    declarations:[
-    ],
-    imports: [
-        CommonModule
-        ],
-    exports: [],
-})
-export class SearchBoxModule {}
