+++ src/app/core/components/search-field/search-field.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { MaterialModule } from 'src/app/material-module';
+import { SearchFieldComponent } from './search-field.component';
+
+@NgModule({
+    declarations:[SearchFieldComponent],
+    imports: [
+        CommonModule,
+        MaterialModule
+        ],
+    exports: [SearchFieldComponent]
+})
+export class SearchFieldModule {}
