+++ src/app/modules/emergency-register/components/add-search-media-dialog/add-search-media-dialog.component.html
@@ -1,18 +1,27 @@
 <h1 mat-dialog-title>Patient search</h1>
 
-<app-search-box (onSearchQuery)="searchQuery")></app-search-box>
+<app-search-field (searchString)="onSearchQuery($event)"></app-search-field>
 
 <mat-accordion>
-    <mat-expansion-panel hideToggle class="expansion-panel-margin"
-    *ngFor="let searchResult of searchResults$ | async"
-    (click)="openMediaDialog(searchResult.TagNumber, searchResult.PatientId)">
-        <mat-expansion-panel-header>
-            <mat-panel-title >
-                {{searchResult.TagNumber}}
-            </mat-panel-title>
-            <mat-panel-description>
-               {{searchResult.EmergencyNumber}}
-            </mat-panel-description>
-        </mat-expansion-panel-header>
-    </mat-expansion-panel>
+
+    <div *ngIf="searchResults$ | async as searchResults;">
+
+        <div *ngIf="searchResults.length === 0;">No results</div>
+
+        <mat-expansion-panel hideToggle class="expansion-panel-margin"
+        *ngFor="let searchResult of searchResults"
+        (click)="openMediaDialog(searchResult.TagNumber, searchResult.PatientId)">
+
+            <mat-expansion-panel-header>
+                <mat-panel-title >
+                    {{searchResult.TagNumber}}
+                </mat-panel-title>
+                <mat-panel-description>
+                    {{searchResult.EmergencyNumber}}
+                </mat-panel-description>
+            </mat-expansion-panel-header>
+        </mat-expansion-panel>
+
+    </div>
+
 </mat-accordion>
