+++ src/app/modules/emergency-register/components/add-search-media-dialog/add-search-media-dialog.module.ts
@@ -1,14 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AddSearchMediaDialogComponent } from './add-search-media-dialog.component';
+import { MaterialModule } from 'src/app/material-module';
+import { SearchFieldModule } from 'src/app/core/components/search-field/search-field.module';
+
 
 @NgModule({
-    declarations:[
-        AddSearchMediaDialogModule,
-    ],
+    declarations:[AddSearchMediaDialogComponent],
     imports: [
         CommonModule,
+        MaterialModule,
+        SearchFieldModule
         ],
-    exports: [AddSearchMediaDialogComponent],
+    exports: [AddSearchMediaDialogComponent]
 })
+
 export class AddSearchMediaDialogModule {}
