+++ src/app/modules/emergency-register/components/outstanding-case-board/outstanding-case-board.component.html
@@ -15,7 +15,7 @@
     </button>
 
   </mat-form-field>
-  
+
   </form>
 
   <div fxLayout="row wrap" fxLayoutAlign="space-between center" fxLayoutGap="20px" class="refresh-box">
@@ -111,8 +111,6 @@
             <div class="name">{{outstandingCase.callerName}}</div>
 
           <mat-icon aria-label="Print emergency record" (click)="printEmergencyCard(outstandingCase.emergencyCaseId)">print</mat-icon>
-
-          
 
             <div fxLayout="row" fxLayoutAlign="start center" >
                 <div class="telephone-dash-container">
