+++ src/app/modules/emergency-register/components/outstanding-case-board/outstanding-case-board.component.ts
@@ -118,7 +118,6 @@ export class OutstandingCaseBoardComponent implements OnInit {
 
     // Attempting to force change detection here causes the whole thing to hang.
     this.outstandingCases$.subscribe((value) => {
-      console.log(value);
         this.loading = false;
         this.changeDetector.detectChanges();
     });
@@ -134,7 +133,7 @@ export class OutstandingCaseBoardComponent implements OnInit {
             patientId,
         }
     });
-    
+
   }
 
   autoRefreshToggled(){
