+++ src/app/modules/emergency-register/components/tab-bar/tab-bar.component.html
@@ -1,5 +1,4 @@
-<mat-tab-group [selectedIndex]="selected.value"
-        (selectedIndexChange)="selected.setValue($event)">
+<mat-tab-group [selectedIndex]="selected.value" (selectedIndexChange)="selected.setValue($event)">
 
         <mat-tab *ngFor="let tab of tabs; let index = index"
         label="tab">
