+++ src/app/modules/emergency-register/pages/emergency-register-page.component.html
@@ -2,4 +2,4 @@
 
 <!-- <input type='file' (change)="onChange($event)" accept=".jpg, .jpeg, .png"> -->
 
-<add-case-media></add-case-media>
+<!--<add-case-media></add-case-media>-->
