+++ src/app/modules/reporting/reporting-page.module.ts
@@ -5,12 +5,14 @@ import { ReportingPageRoutingModule } from './reporting-page-routing.module';
 import { SurgeriesByDateDialogComponent } from './components/surgeries-by-date-dialog/surgeries-by-date-dialog.component';
 import { ReportingPageComponent } from './pages/reporting-page/reporting-page.component';
 import { PatientDetailsDialogComponent } from './components/patient-details-dialog/patient-details-dialog.component';
+import { TreatmentRecordComponent } from 'src/app/core/components/treatment-record/treatment-record.component';
 
 @NgModule({
     declarations: [
         ReportingPageComponent,
         SurgeriesByDateDialogComponent,
-        PatientDetailsDialogComponent
+        PatientDetailsDialogComponent,
+        TreatmentRecordComponent
     ],
     imports: [
         CommonModule,
