+++ src/manifest.webmanifest
@@ -12,9 +12,6 @@
     "method": "POST",
     "enctype": "multipart/form-data",
     "params": {
-      "title": "name",
-      "text": "description",
-      "url": "link",
       "files": [
         {
           "name": "image",
